Updates to temp. correct impl of waste metric

+++ src/algorithms/lowestlarger.rs
@@ -53,7 +53,7 @@ pub fn select_coin_lowestlarger(
     if accumulated_value < (target + estimated_fees) {
         Err(SelectionError::InsufficientFunds)
     } else {
-        let waste: u64 = calculate_waste(
+        let waste: f32 = calculate_waste(
             options,
             accumulated_value,
             accumulated_weight,
